Allow different class names in rust and Godot
There's an issue that two structs with the same name in different
modules will conflict with each other when sent to Godot since only the
struct name is considered.  This resolves that by allowing the API user
to manually change the name of the class as Godot understands it.  This
also improves the error message that occurs when there are aliased
classes.

Future work may seek to catch this issue at compile time rather than at
runtime.