CI: update LLVM from 10.0 to 15.0.7 -> fix spurious macOS failures
Fixes an issue where macOS CI spuriously failed, because it could not read the Godot version (or invoke the
Rust Command API on Godot in general). Trying to look for the issue in all sorts of dimensions (Godot version,
rustc version, gdext's own config changes), it turned out that LLVM was the only thing that could sustainably
address it. Which is also a bit strange, since v10.0 has not caused issues before, only since around start of
October 2023. If it reappears, one more thing to look into would be versions of GitHub actions (that are not
locked, but auto-updated).